project-status-over-time-etl/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Faker==24.8.0